+++ NScumm.Core/Graphics/ClassicCostumeRenderer.cs
@@ -278,8 +278,8 @@ namespace NScumm.Core.Graphics
                     {
                         _width = _loaded.CostumeReader.ReadByte() * 8;
                         _height = _loaded.CostumeReader.ReadByte();
-                        xmoveCur = _xmove + _loaded.CostumeReader.ReadByte() * 8;
-                        ymoveCur = _ymove - _loaded.CostumeReader.ReadByte();
+                        xmoveCur = _xmove + (sbyte)_loaded.CostumeReader.ReadByte() * 8;
+                        ymoveCur = _ymove - (sbyte)_loaded.CostumeReader.ReadByte();
                         _xmove += (sbyte)_loaded.CostumeReader.ReadByte() * 8;
                         _ymove -= (sbyte)_loaded.CostumeReader.ReadByte();
                         _srcptr += 6;
